refactor: Rename nvda_patcher.py to extensionMapper.py and clean up code
